+++ src/abstract/abstract_driver.mli
@@ -1,8 +1,0 @@
-(* SPDX-License-Identifier: AGPL-3.0-or-later *)
-(* Copyright © 2021-2026 OCamlPro *)
-(* Written by the Owi programmers *)
-
-val modul :
-     Abstract_extern.Func.t Link.State.t
-  -> Abstract_extern.Func.t Linked.Module.t
-  -> Abstract_state.t