+++ src/abstract/abstract_f64.ml
@@ -11,3 +11,7 @@ let unknown ctx = Abstract_domain.binary_unknown ~size ctx
 let of_float ctx _ = unknown ctx
 
 let to_binary f = f
+
+let eq ctx f1 f2 = Abstract_domain.Binary_Forward.beq ~size ctx f1 f2
+
+let reinterpret_i64 ctx _i = unknown ctx